Refactored EventOwner to EventOrganizer
